Present revised schedule only when there is a tangible difference
A tangible difference being at least a minute because there seemed to be
a lot of estimates which were actually the same as scheduled.

Extracted the data building data into its own class of pure functions,
which contains the schedule extractor. This makes for easier testing of
the logic in these functions and allows the view model to build and
store the data objects so they are not being rebuilt on each request.
This will also allow widgets to selectively rebuild based on these data
objects changing, to be implemented in a following change.